// src/core/enums.ts
export enum Raza {
  Astrales = "Astrales",
  Celidos = "Célidos",
  Nebulanos = "Nebulanos",
}

export enum Genero {
  Masculino = "Masculino",
  Femenino = "Femenino",
}

export enum RarezaSkin {
  Basico = "Básico",
  Epico = "Épico",
  Collector = "Collector",
}

export enum EstadoAula {
  Estudiando = "Estudiando",
  Descanso = "Descanso",
}

export enum RolHeroe {
  GuardianDelEnfoque = "GuardianDelEnfoque",
  EspecialistaAcademico = "EspecialistaAcademico",
  SoporteMotivacional = "SoporteMotivacional",
}